Add student table,contrller and create method

+++ Database.Domain/Interfaces/IGenericRepository.cs
@@ -9,7 +9,7 @@ namespace Database.Domain.Interfaces
     public interface IGenericRepository<T> where T : class
     {
         IEnumerable<T> GetAll();
-        T GetByID(int id);
+        T GetByID(long id);
         void Add(T entity);
         void Update(T entity);
         void Delete(T entity);
